+++ app/permissions/page.tsx
@@ -1,0 +1,7 @@
+"use client";
+
+import PermissionsGraph from "@/components/permissions/PermissionsGraph";
+
+export default function PermissionsPage() {
+  return <PermissionsGraph />;
+}